style(module1)
All functions related to the initials screens were created.
