+++ Makefile.win
@@ -4,11 +4,11 @@
 CPP      = g++.exe
 CC       = gcc.exe
 WINDRES  = windres.exe
-OBJ      = src/main.o src/SDLfunctions.o src/selectionScreens.o src/cJSON.o src/cJSON_Utils.o
-LINKOBJ  = src/main.o src/SDLfunctions.o src/selectionScreens.o src/cJSON.o src/cJSON_Utils.o
-LIBS     = -L"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib" -L"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/lib" -L"./lib" -static-libgcc -L"./lib" -lmingw32 -lSDL2main -lSDL2 -lSDL2_image -lSDL2_ttf
-INCS     = -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"./include" -I"./include"
-CXXINCS  = -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include/c++" -I"./include"
+OBJ      = src/main.o src/SDLfunctions.o src/selectionScreens.o src/cJSON.o src/cJSON_Utils.o readFiles.o
+LINKOBJ  = src/main.o src/SDLfunctions.o src/selectionScreens.o src/cJSON.o src/cJSON_Utils.o readFiles.o
+LIBS     = -L"D:/Dev-Cpp/MinGW64/lib" -L"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/lib" -L"./lib" -static-libgcc -L"./lib" -lmingw32  -lSDL2main -lSDL2  -lSDL2_image  -lSDL2_ttf
+INCS     = -I"D:/Dev-Cpp/MinGW64/include" -I"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"./include" -I"./include"
+CXXINCS  = -I"D:/Dev-Cpp/MinGW64/include" -I"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include/c++" -I"./include"
 BIN      = EchoesOfHonor.exe
 CXXFLAGS = $(CXXINCS) 
 CFLAGS   = $(INCS) 
@@ -38,3 +38,6 @@ src/cJSON.o: src/cJSON.c
 
 src/cJSON_Utils.o: src/cJSON_Utils.c
 	$(CC) -c src/cJSON_Utils.c -o src/cJSON_Utils.o $(CFLAGS)
+
+readFiles.o: readFiles.c
+	$(CC) -c readFiles.c -o readFiles.o $(CFLAGS)
