+++ include/SDLfunctions.h
@@ -11,6 +11,6 @@ tile_t* loadImage(int numImage, char *path, SDL_Renderer *renderer);
 
 void defineSize(int numTile, tile_t *tileset, int w, int h);
 
-void createText(char *fontPath, int fontSize, int *color, SDL_Renderer *renderer, char *text);
+void createText(char *fontPath, int fontSize, int *color, SDL_Renderer *renderer, char *text, tile_t *textTile);
 
 #endif
