+++ maps/Map1.json
@@ -1,0 +1,10 @@
+{
+"phase": 1,
+"health": 28,
+"speed": 34,
+"endurance": 21,
+"damage": 32,
+"xp": 0,
+"level": 1,
+"character": "Samurai"
+}
