+++ maps/Map3.json
@@ -1,0 +1,10 @@
+{
+    "phase": 1,
+    "health": 0,
+    "speed": 0,
+    "endurance": 0,
+    "damage": 0,
+    "xp": 0,
+    "level": 32,
+    "character": "Archer"
+}
